# server/src/swagger.json
{
	"swagger": "2.0",
	"info": {
		"version": "1.0.0",
		"title": "Mern Stack Boilerplate"
	},
	"tags": [
		{
			"name": "Health",
			"description": "API for system health"
		},
		{
			"name": "Users",
			"description": "API for users in the system"
		}
	],
	"consumes": [
		"application/json"
	],
	"produces": [
		"application/json"
	],
	"paths": {
		"/api/health": {
			"get": {
				"tags": [
					"Health"
				],
				"summary": "Get system health",
				"responses": {
					"200": {
						"description": "OK",
						"type": "string"
					}
				}
			}
		},
		"/api/users": {
			"get": {
				"tags": [
					"Users"
				],
				"summary": "Get all users in system",
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/Users"
						}
					}
				}
			},
			"post": {
				"tags": [
					"Users"
				],
				"summary": "Create new user in system",
				"parameters": [
					{
						"name": "user",
						"in": "body",
						"description": "User that we want to create",
						"required": true,
						"schema": {
							"$ref": "#/definitions/User"
						}
					}
				],
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "New user is created",
						"schema": {
							"$ref": "#/definitions/User"
						}
					}
				}
			}
		},
		"/api/users/{id}": {
			"get": {
				"tags": [
					"Users"
				],
				"summary": "Get specific user in system",
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/Users"
						}
					}
				},
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"description": "Id of the user",
						"type": "string"
					}
				]
			},
			"patch": {
				"tags": [
					"Users"
				],
				"summary": "Update user",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"description": "Id of the user",
						"type": "string"
					},
					{
						"name": "user",
						"in": "body",
						"description": "User that we want to create",
						"required": true,
						"schema": {
							"$ref": "#/definitions/User"
						}
					}
				],
				"responses": {
					"200": {
						"description": "User is updated",
						"schema": {
							"$ref": "#/definitions/User"
						}
					}
				}
			},
			"delete": {
				"tags": [
					"Users"
				],
				"summary": "Get specific user in system",
				"responses": {
					"200": {
						"description": "OK"
					}
				},
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"description": "Id of the user",
						"type": "string"
					}
				]
			}
		}
	},
	"definitions": {
		"User": {
			"required": [
				"firstName",
				"lastName"
			],
			"properties": {
				"_id": {
					"type": "string",
					"uniqueItems": true
				},
				"firstName": {
					"type": "string"
				},
				"lastName": {
					"type": "string"
				},
				"username": {
					"type": "string"
				},
				"password": {
					"type": "string"
				}
			}
		},
		"Users": {
			"type": "array",
			"$ref": "#/definitions/User"
		}
	}
}